Add files via upload
In this shopping list I have used the DOM callback functions. User can add the items to the shopping list and remove the item from the list. When user hit enter or click enter the item will add to the end of the list and when user click delete it will delete the item from the top of the list.